Extract interface from LatestCompleteStateNexus
