+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/cli/DiagramCommand.java
@@ -35,7 +35,7 @@ import com.swirlds.config.api.Configuration;
 import com.swirlds.platform.config.DefaultConfiguration;
 import com.swirlds.platform.config.StateConfig;
 import com.swirlds.platform.eventhandling.TransactionPool;
-import com.swirlds.platform.state.nexus.LatestCompleteStateNexus;
+import com.swirlds.platform.state.nexus.DefaultLatestCompleteStateNexus;
 import com.swirlds.platform.system.status.PlatformStatusManager;
 import com.swirlds.platform.wiring.PlatformWiring;
 import edu.umd.cs.findbugs.annotations.NonNull;
@@ -106,7 +106,7 @@ public final class DiagramCommand extends AbstractCommand {
         platformWiring.wireExternalComponents(
                 new PlatformStatusManager(platformContext, platformContext.getTime(), threadManager, a -> {}),
                 new TransactionPool(platformContext),
-                new LatestCompleteStateNexus(
+                new DefaultLatestCompleteStateNexus(
                         platformContext.getConfiguration().getConfigData(StateConfig.class),
                         platformContext.getMetrics()),
                 notificationEngine);
